home(visionOS): fix crash
